+++ baybe/recommenders/pure/__init__.py
@@ -4,7 +4,7 @@ Pure recommenders implement selection algorithms and can be queried for providin
 recommendations. They can be part of meta recommenders.
 """
 
-from baybe.recommenders.pure.bayesian.botorch import BotorchRecommender
+from baybe.recommenders.pure.bayesian import BotorchRecommender
 from baybe.recommenders.pure.nonpredictive import (
     FPSRecommender,
     GaussianMixtureClusteringRecommender,
@@ -12,9 +12,11 @@ from baybe.recommenders.pure.nonpredictive import (
     PAMClusteringRecommender,
     RandomRecommender,
 )
+from baybe.recommenders.pure.tfpr import TFPRRecommender
 
 __all__ = [
     "BotorchRecommender",
+    "TFPRRecommender",
     "FPSRecommender",
     "GaussianMixtureClusteringRecommender",
     "KMeansClusteringRecommender",